+++ Changes.md
@@ -1,5 +1,7 @@
 WIP  TBD
 
+ * Rewrote TestLoaderSaver so that loader fails before saver runs and also provides a list of Buffers once saver's writer is closed. This behaves as a better analog of the file system loader/saver as well.
+ * Added a test for the test loader/saver.
  * Improved testing of filesystem loader/saver.
 
 v0.0.1  2023-07-28